[1772]: Move the helpers package into internal and split it up a bit. The exit code went into the command stuff since that's all that used it.
