+++ x/exchange/market_test.go
@@ -13,7 +13,7 @@ import (
 	sdkmath "cosmossdk.io/math"
 	sdk "github.com/cosmos/cosmos-sdk/types"
 
-	"github.com/provenance-io/provenance/helpers"
+	"github.com/provenance-io/provenance/internal/helpers"
 	"github.com/provenance-io/provenance/testutil/assertions"
 )
